Debugger was breaking because buffer wasn't clearing on printf()
To fix this, every printf statement must terminate with a new line character. I.e. printf("blah blah \n"); 

The reason the parser won't work after the first input is because of an anomalous new line character that appears and prepends itself to the user input. So, instead of having "add" it will store "[0x0a]add".